fix: remove unnecessary cast in loyalty rule mapper

+++ heatmap-api/src/main/java/ru/tbank/tmap/loyalty/presentation/mapper/BusinessLoyaltyRuleMapper.java
@@ -37,7 +37,7 @@ public class BusinessLoyaltyRuleMapper {
 
     public LoyaltyRuleResponse toResponse(final LoyaltyRuleDetails details) {
         final LoyaltyRule rule = details.rule();
-        final long remainingUsages = Math.max(0L, (long) rule.getMaxUsages() - details.currentUsages());
+        final long remainingUsages = Math.max(0L, rule.getMaxUsages() - details.currentUsages());
         return new LoyaltyRuleResponse()
                 .id(rule.getId())
                 .venueId(rule.getVenueId())
